+++ src/services/CredentialsService.test.js
@@ -1,70 +1,70 @@
 // Third-party dependencies
-import chai, { expect } from 'chai'
-import sinonChai from 'sinon-chai'
-import {
-    beforeEach,
-    describe,
-    it,
-} from 'mocha'
+import { getUniqueId } from 'react-native-device-info'
 
 // In-house dependencies
 import * as CredentialsService from './CredentialsService'
 
-// Setup Chai
-chai.use(sinonChai)
+// Setup mocks for this whole test file
+getUniqueId.mockReturnValue('DEVICE_ID')
 
 describe('CredentialsService', () => {
+    let testContext
+
+    beforeEach(() => {
+        testContext = {}
+    })
+
     describe('getApiKey', () => {
         it('at this point, should just return the device ID', () => {
             const actualApiKey = CredentialsService.getApiKey()
 
-            expect(actualApiKey).to.equal('DEVICE_ID')
+            expect(actualApiKey).toBe('DEVICE_ID')
         })
     })
 
     describe('sanitizeApiKey', () => {
         beforeEach(() => {
-            this.apiKey = CredentialsService.getApiKey()
+            testContext.apiKey = CredentialsService.getApiKey()
         })
 
         it('replaces the device ID at the beginning of the string', () => {
-            const input = `${this.apiKey}andmanycharactersafterwards`
+            const input = `${testContext.apiKey}andmanycharactersafterwards`
 
             const sanitizedString = CredentialsService.sanitizeApiKey(input)
 
-            expect(sanitizedString).to.equal('SANITIZED_API_KEYandmanycharactersafterwards')
+            expect(sanitizedString).toBe('SANITIZED_API_KEYandmanycharactersafterwards')
         })
 
         it('replaces the device ID in the middle of the string', () => {
-            const input = `manycharactersbefore${this.apiKey}andmanycharactersafterwards`
+            const input = `manycharactersbefore${testContext.apiKey}andmanycharactersafterwards`
 
             const sanitizedString = CredentialsService.sanitizeApiKey(input)
 
-            expect(sanitizedString).to.equal('manycharactersbeforeSANITIZED_API_KEYandmanycharactersafterwards')
+            expect(sanitizedString).toBe('manycharactersbeforeSANITIZED_API_KEYandmanycharactersafterwards')
         })
 
         it('replaces the device ID at the end of the string', () => {
-            const input = `manycharactersbefore${this.apiKey}`
+            const input = `manycharactersbefore${testContext.apiKey}`
 
             const sanitizedString = CredentialsService.sanitizeApiKey(input)
 
-            expect(sanitizedString).to.equal('manycharactersbeforeSANITIZED_API_KEY')
+            expect(sanitizedString).toBe('manycharactersbeforeSANITIZED_API_KEY')
         })
 
         it('replaces the device ID more than once', () => {
-            const input = `${this.apiKey}and${this.apiKey}more${this.apiKey}`
+            const input = `${testContext.apiKey}and${testContext.apiKey}more${testContext.apiKey}`
 
             const sanitizedString = CredentialsService.sanitizeApiKey(input)
 
-            expect(sanitizedString).to.equal('SANITIZED_API_KEYandSANITIZED_API_KEYmoreSANITIZED_API_KEY')
+            expect(sanitizedString).toBe('SANITIZED_API_KEYandSANITIZED_API_KEYmoreSANITIZED_API_KEY')
         })
         
         it('does not replace anything if the device ID is not in the string', () => {
-            const input = `before${this.apiKey.substring(1)}after`    // The API key without its first character
+            const input = `before${testContext.apiKey.substring(1)}after`    // The API key without its first character
 
             const sanitizedString = CredentialsService.sanitizeApiKey(input)
 
-            expect(sanitizedString).to.equal('beforeEVICE_IDafter')
+            expect(sanitizedString).toBe('beforeEVICE_IDafter')
         })
     })
 })