+++ src/services/SensorApiService.test.js
@@ -1,58 +1,45 @@
-// Third-party dependencies
-import chai, { expect } from 'chai'
-import sinon from 'sinon'
-import sinonChai from 'sinon-chai'
-import {
-    afterEach,
-    beforeEach,
-    describe,
-    it,
-} from 'mocha'
-
-chai.use(sinonChai)
-
 // In-house dependencies
 import SensorApiService from './SensorApiService'
 import * as FetchService from './FetchService'
 
+// Setup mocks for this whole test file
+jest.mock('./FetchService', () => ({
+    post: jest.fn()
+}))
+jest.mock('./CredentialsService', () => ({
+    getApiKey: jest.fn().mockReturnValue('API_KEY')
+}))
+
 describe('SensorApiService', () => {
-    beforeEach(() => {
-        sinon.stub(FetchService, 'post')
-    })
-
-    afterEach(() => {
-        FetchService.post.restore()
-    })
-
     describe('testRequest', () => {
-        it('calls POST with the correct URI', () => {
+        it('calls POST with the correct URI, base URL, and headers', () => {
             const expectedParameters = {
                 uri: '/alert/test',
                 base: 'https://sensor.com',
                 headers: {
-                    'X-API-KEY': 'DEVICE_ID'
+                    'X-API-KEY': 'API_KEY'
                 },
             }
 
             SensorApiService.testRequest()
 
-            expect(FetchService.post).to.have.been.calledWith(expectedParameters)
+            expect(FetchService.post).toHaveBeenCalledWith(expectedParameters)
         })
     })
 
     describe('fakeEndpointRequest', () => {
-        it('calls POST with the correct URI', () => {
+        it('calls POST with the correct URI, base URL, and headers', () => {
             const expectedParameters = {
                 uri: 'not/in/real/life',
                 base: 'https://sensor.com',
                 headers: {
-                    'X-API-KEY': 'DEVICE_ID'
+                    'X-API-KEY': 'API_KEY'
                 },
             }
 
             SensorApiService.fakeEndpointRequest()
 
-            expect(FetchService.post).to.have.been.calledWith(expectedParameters)
+            expect(FetchService.post).toHaveBeenCalledWith(expectedParameters)
         })
     })
 })