+++ CHANGELOG.md
@@ -23,14 +23,13 @@ the code was deployed.
 - Redux for handling state (CU-d8qufg).
 - An ErrorReportingServie (CU-d8qufg).
 - API calls (CU-d8qufg).
-- Mocha/Chai/Sinon test framework and tests (CU-d8qufg).
 - Standard Logger class (CU-d8qufg).
 - SoundService to be used for alert and message noises (CU-d8qufg).
 - CredentialsService to handle the Brave Alert API's API key (CU-d8qufg).
+- Services tests (CU-d8qufg).
 
 ### Removed
 - `yarn.lock` (CU-d8qufg).
-- Jest testing library and example test (CU-d8qufg).
 
 ### Fixed
 - iOS deployment target (CU-d8qufg).